added the stats for the classroom
