+++ src/components/pages/classroom/Classroom.tsx
@@ -2,13 +2,13 @@ import { memo } from 'react';
 import AsideAdmin from 'components/layout/aside-admin/AsideAdmin';
 import Hero from './hero/Hero';
 import LiveClasses from './live-classes/LiveClasses';
-import TopStats from './top-stats/TopStats';
+import ClassroomStats from './classroom-stats/ClassroomStats';
 
-const Classrooms = () => (
+const Classroom = () => (
   <section className="flex _full flex-wrap max-h-full overflow-auto lg:flex-nowrap gap-6">
     <div className="w-full lg:w-3/5 xl:w-8/12 shrink-0 overflow-auto">
       <Hero />
-      <TopStats />
+      <ClassroomStats />
       <LiveClasses />
     </div>
     <div className="w-full lg:w-auto grow overflow-auto">
@@ -17,4 +17,4 @@ const Classrooms = () => (
   </section>
 );
 
-export default memo(Classrooms);
+export default memo(Classroom);
